Clean code: Rename TransferTable.java to TransferService and move it to the transfer package.
