+++ src/main/java/org/opentripplanner/routing/graph/Graph.java
@@ -72,10 +72,10 @@ import org.opentripplanner.model.calendar.CalendarServiceData;
 import org.opentripplanner.model.calendar.ServiceDate;
 import org.opentripplanner.model.calendar.impl.CalendarServiceImpl;
 import org.opentripplanner.model.projectinfo.OtpProjectInfo;
+import org.opentripplanner.model.transfer.TransferService;
 import org.opentripplanner.routing.algorithm.raptor.transit.TransitLayer;
 import org.opentripplanner.routing.algorithm.raptor.transit.mappers.TransitLayerUpdater;
 import org.opentripplanner.routing.bike_rental.BikeRentalStationService;
-import org.opentripplanner.routing.core.TransferTable;
 import org.opentripplanner.routing.edgetype.EdgeWithCleanup;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.impl.DelegatingTransitAlertServiceImpl;
@@ -123,7 +123,7 @@ public class Graph implements Serializable {
 
     private final Map<Class<?>, Serializable> services = new HashMap<>();
 
-    private final TransferTable transferTable = new TransferTable();
+    private final TransferService transferService = new TransferService();
 
     private GraphBundle bundle;
 
@@ -522,8 +522,8 @@ public class Graph implements Serializable {
         return env;
     }
 
-    public TransferTable getTransferTable() {
-        return transferTable;
+    public TransferService getTransferTable() {
+        return transferService;
     }
 
     // Infer the time period covered by the transit feed
